DRBayes: Bayesian Doubly Robust Causal Inference Methods

Bayesian doubly robust estimation of average treatment effects via posterior coupling, in which the outcome-model and propensity-score posteriors are linked so that the doubly robust moment condition is satisfied. Gibbs samplers for linear, logistic and probit models under conjugate or horseshoe priors are provided, and posterior draws produced by external Bayesian software can be supplied instead. A Bayesian bootstrap estimator and utilities for replicated simulation studies are also included. The methodology is described in Orihara, Momozaki and Sugasawa (2025) doi:10.48550/arXiv.2506.04868 .

See also

Author

Maintainer: Tomotaka Momozaki t_momozaki@rs.tus.ac.jp

Authors:

Tomotaka Momozaki t_momozaki@rs.tus.ac.jp

Shunichiro Orihara

Shonosuke Sugasawa